+++ application/Sourcecode/Player/Attack/AttackAir2.cpp
@@ -82,7 +82,7 @@ void AttackAir2::MoveUpdate()
 	
 
 	spline_.Update();
-	attackCol_.at(0)->damageCoolTime = timer_.GetLimitTimer() - timer_.GetTimer();
+	attackCol_.at(0)->damageCoolTime = (uint32_t)timer_.GetLimitTimer() - (int32_t)timer_.GetTimer();
 	attackCol_.at(0)->col_.center = spline_.GetNowPoint();
 	swordPos_ = attackCol_.at(0)->col_.center;
 }
